Dns Enumerator/requirements.txt
dnspython==2.2.1
Pillow==9.3.0
reportlab==3.6.12
